Support questions where answers are neither right nor wrong
The xAPI specification allows questions where answers are
neither right nor wrong. In that case, the correct responses
pattern is omitted, cmp. https://github.com/adlnet/xAPI-Spec/blob/master/xAPI-Data.md#correct-responses-pattern
Currently, this is interpreted wrong.

With these changes, if the xAPI statement does not contain a
correct responses pattern for an xAPI choice interaction, it
will use a blue checkmark for indicating that the answer was given,
relabel the "Answers" column to "Chosen" and suppress the "Correct
Answers" column.

The behavior for xAPI choice interactions that provide a correct
responses pattern is not changed, so existing content types such
as the multiple choice quiz are not affected.