Support questions where answers are neither right nor wrong
The xAPI specification allows questions where answers are
neither right nor wrong. In that case, the correct responses
pattern is omitted, cmp. https://github.com/adlnet/xAPI-Spec/blob/master/xAPI-Data.md#correct-responses-pattern
Currently, this is interpreted wrong.

With these changes, if the xAPI statement does not contain a
correct responses pattern for an xAPI fill-in interaction, it
will display the response without marking the fields that were
filled in in any way. It will also suppress the footer with the
explanations for markings.

The behavior for xAPI fill-in interactions that provide a correct
responses pattern is not changed, so existing content types such
as the fill in the blanks or mark the words are not affected.